+++ src/RPJ.hpp
@@ -21,3 +21,5 @@ extern Model *modelCircularRide;
 extern Model *modelDrillingHoles;
 extern Model *modelTuxOn;
 extern Model *modelPigeonPlink;
+extern Model *modelGenie;
+extern Model *modelGenieExpander;