+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Disable editing tooltips of Mode knob in Grendel
+- Genie sync/trigger
+- Genie length knobs and cv's
 
 ## [2.1.2] - 2022-06-16
 - Fixed speed bug in TuxOn